requirements.txt
streamlit==1.33.0
langchain==0.1.17
langchain-community==0.0.36
openai==1.24.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.1
transformers==4.40.1
torch==2.2.2
sentencepiece==0.2.0
protobuf==4.25.3
huggingface-hub==0.23.0
sentence-transformers==2.7.0
